refactor all directories under app-components to lowercase and rename component files to uppercase
