+++ src/app-components/CirclePics.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { isValidArrWithValues } from '../../functions';
+import { isValidArrWithValues } from '../functions';
 
 const imgStyle = {
 	height: '12rem',
