+++ src/app-components/LocationDetail/Header.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from 'prop-types';
-import "../location-detail.css";
+import "./location-detail.css";
 
 const LocationDetailHeader = ( { onExpand } ) => {
