+++ src/app-components/SearchBox.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from "classnames";
-import './searchbox.css';
 
 const SearchBox = ({value,onChange,onEnterKey,text}) => {
